\name{chisq.in}
\alias{chisq.in}
\title{This is the function to calculate the chi-square in statistics in COP 
}
\description{
The function is used to calculate the Chi-square in statistics in COP
}
\usage{
chisq.in(lambdaf, lambdar, n)
}
\arguments{
  \item{lambdaf}{ the lambda for the full model
}
  \item{lambdar}{the lambda for the reduced model
}
  \item{n}{the number of observations
}
}
\value{
}
\references{
}
\author{
Wenxuan Zhong
}
\note{
}


\seealso{}
\examples{}
