\name{chisq.out}
\alias{chisq.out}

\title{
The function is used to calculate chisquare out statistics in COP
}
\description{
The function is used to calculate the COP statistics for backward delete variables}
\usage{
chisq.out(lambdaf, lambdar, n)
}
\arguments{
  \item{lambdaf}{the lambda for the full model
}
  \item{lambdar}{the lambda for the reduced model
}
  \item{n}{number of observations
}
}
\value{

}
\references{}

\author{
Wenxuan Zhong
}
\note{

}



\seealso{
}
\examples{
}
