\name{COP}
\alias{COP-package}
\alias{COP}
\docType{package}
\title{
This package is for selecting variables for index models using correlation pursuit.
}
\description{This packages is for selecting variables for index models using correlation pursuit method. 
Correlation pursuit (COP) can be viewed as a
generalization of the conventional linear stepwise regression method 
to semi-parametric regression models. Unlike the
conventional stepwise, COP selects variables that maximize the
correlation between a transformed response and a linear
combination of the predictors. A sequential selection strategy is
used to select variables on multiple linear combinations.}
\details{
\tabular{ll}{
Package: \tab COP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-17\cr
License: \tab \cr
}
}
\author{
Wenxuan Zhong	

Maintainer: Wenxuan Zhong <wenxuan@illinois.edu>
}
\references{
Correlation Pursuit: Stepwise Variable Selection for Index Models
}
\seealso{
}
\examples{
}
