\name{forward.corp}
\alias{forward.corp}
\title{
Forward selection in COP
}
\description{
This function is for forward selection in COP
}
\usage{
forward.corp(x, y, ind, H, signif.in, dd, lambdar)
}
\arguments{
  \item{x}{is a n by p matrix for the covariates
}
  \item{y}{
is a response variable
}
  \item{ind}{
is the subset of variable that has not been selected at previous step
}
  \item{H}{
is the number of slices
}
  \item{signif.in}{
is the threshold to add a variable in the selected subset of variables
}
  \item{dd}{is the direction
}
  \item{lambdar}{
the critical value based on the reduced model
}
}

\value{

}
\references{

}
\author{
Wenxuan Zhong
}




\seealso{

}
\examples{
}
