% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{mpi_print}
\alias{mpi_print}
\alias{mpi_cat}
\title{MPI Printer}
\usage{
mpi_cat(x, quiet = FALSE, all.rank = FALSE, comm = 0L)

mpi_print(x, quiet = FALSE, all.rank = FALSE, comm = 0L)
}
\arguments{
\item{x}{The input to be printed.}

\item{quiet}{Should ranks print their rank number before printing the input?}

\item{all.rank}{Should all ranks print or only rank 0?}

\item{comm}{MPI communicator number.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Orderly printing of data across MPI ranks. Similar to
\code{pbdMPI::comm.print()}.
}
\details{
All printing is done on rank 0 to ensure printing occurs in rank order. This
requires communication if \code{all.rank=TRUE}.
}
\examples{
\dontrun{
suppressMessages(library(cop))

x = matrix(1:4, 2) + 10*comm.rank()
mpi_print(x)

finalize()
}

}
