% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_gen.r
\name{sparse_gen}
\alias{sparse_gen}
\alias{spar_gen_bandish}
\alias{spar_gen_banded}
\title{Sparse Matrix Generators}
\usage{
spar_gen_bandish(nrows, ncols, seed)

spar_gen_banded(nrows, ncols, band = 2)
}
\arguments{
\item{nrows, ncols}{The dimension}

\item{seed}{Random seed.}

\item{band}{The band size, a positive integer. \code{band=1} returns a matrix whose
diagonal entries are all 1; \code{band=2} returns a matrix with 1 on each
diagonal entry and 1 on each entry directly above and below the main
diagonal; etc.}
}
\value{
A sparse \code{"dgCMatrix"} object is returned.
}
\description{
Generate sparse matrices.
}
\details{
\code{spar_gen_bandish()} generates a "band-ish" matrix, a matrix whose
entries are probabilistically more likely to be 1 the closer to the main
diagonal they are.

\code{spar_gen_banded()} generates a banded matrix.
}
