\name{starting}
\alias{starting}

\title{
starting is a function to select the first two important predictors to start with}
\description{This function is used to select the first one or two predictors or check whether there is any predictor significant.
}
\usage{
starting(x, y, H)
}
\arguments{
  \item{x}{is a n by p matrix for the covariates
}
  \item{y}{
is a response variable
}
  \item{H}{
is the number of slices
}
}

\value{

}
\references{

}
\author{
Wenxuan Zhong
}

\seealso{}
\examples{}

