% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop-package.r
\docType{package}
\name{cop-package}
\alias{cop-package}
\title{Custom OPs}
\description{
Some custom reduction operations for data distributed over MPI.
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
