% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadd_allreduce.r
\name{spadd_allreduce}
\alias{spadd_allreduce}
\alias{spadd_reduce}
\alias{spadd_allreduce.dgCMatrix}
\alias{spadd_reduce.dgCMatrix}
\title{Sparse Adder Reduction}
\usage{
spadd_allreduce(x, comm = 0L, type = "double")

spadd_reduce(x, root = 0L, comm = 0L, type = "double")

\method{spadd_allreduce}{dgCMatrix}(x, comm = 0L, type = "double")

\method{spadd_reduce}{dgCMatrix}(x, root = 0L, comm = 0L, type = "double")
}
\arguments{
\item{x}{The input data. Should be a sparse \code{"dgCMatrix"} object. The matrix
should be the same dimensions across all processes.}

\item{comm}{MPI communicator number.}

\item{type}{The type used for the intermediate operations. Should be one of
"double" or "float".}

\item{root}{MPI rank that should receive the return in the non-all version.}
}
\value{
If the all version is called or if the calling rank is equal to \code{root},
then a sparse \code{"dgCMatrix"} object is returned, and otherwise
\code{NULL}.
}
\description{
Given a sparse matrix on each process, this computes the sum of all of those
matrices. The return will be stored as a sparse matrix, although it is not
guaranteed to be sparse.
}
